fix:updated refetching and reassinning to user problem fixed
