+++ frontend/hooks/instructor/useCreateInstructor.ts
@@ -1,21 +1,15 @@
 import { useMutation } from "@tanstack/react-query";
 import { createInstructor } from "@/api/instructor";
 import { InstructorFormData } from "@/types/instructor";
-import { useAuthStore } from "@/stores/auth.store";
 import { User } from "@/types/user";
 import { toast } from "sonner";
-import { queryClient } from "@/utils/queryClient";
-import { ApiResponse } from "@/types/apiResponse";
+import { ApiResponse } from "@/types/general";
 
 export function useCreateInstructor() {
-	const { setUser } = useAuthStore();
 
 	return useMutation<ApiResponse<User>, any, InstructorFormData>({
 		mutationFn: (data: InstructorFormData) => createInstructor(data),
-		onSuccess: (response) => {
-			setUser(response.data);
-			queryClient.invalidateQueries({ queryKey: ["user"], exact: false });
-			queryClient.invalidateQueries({ queryKey: ["instructor"], exact: false });
+		onSuccess: () => {
 			toast.success("Instructor application submitted successfully!", {
 				description: "Your application is under review.",
 			});
